Refactor global positioning tracing into tracer
