+++ src/colmap/estimators/CMakeLists.txt
@@ -44,6 +44,7 @@ COLMAP_ADD_LIBRARY(
         generalized_pose.h generalized_pose.cc
         global_positioning.h global_positioning.cc
         global_positioning_trace.h global_positioning_trace.cc
+        global_positioning_tracer.h global_positioning_tracer.cc
         gravity_refinement.h gravity_refinement.cc
         pose.h pose.cc
         rotation_averaging.h rotation_averaging.cc
